requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
bcrypt==4.1.2
python-dotenv==1.0.0
email-validator==2.1.0
WTForms==3.1.2
reportlab==4.0.8
Flask-Mail==0.9.1